Enable dispersion for paths that encounter dispersive materials
When a path hits a dispersive material (so far only transmissive ones)
a wavelength is sampled (currently uniform on the visible domain). The
material's IOR is recalculated for this wavelength. The radiance
function then returns a result with only the intensity for that
wavelength instead of the whole spectrum. This means that until a lot of
samples are taken dispersive materials look darker than they should be.