Improve acceleration structure selection
Instead of a list of structures with indexes into it we make a struct to
hold Options of each acceleration structure and use an enum to select
which acceleration stucture we want.

This also allows us to only construct the structure we want to use
saving construction time and memory.