Fix segfault upon closing the preview app
A segfault occured because of a bug in luminance where an instance of a
struct containing a refererence to GPU data is dropped after the
context. Luminance does tries to free the GPU data even when the GL
context no longer exists leading to a segfault.

https://github.com/phaazon/luminance-rs/issues/304

This commit works around the bug by not moving the context into the
closure, causing it to be dropped later.