Do not compute or store visibility data for empty voxels
This enables scaling to higher resolutions without as much overhead
since when the voxels get smaller more of them will be empty.

To simplify saving the new hashmaps in scene statistics to a file I am
using serde with postcard to simply serialize the entire struct to a
compact binary format.