Replace rand's thread_rng with a custom one based on rand's SmallRng
This allows us to manually seed the rng, and doing this per path
based on a hash of the pixel coordinate and sample number enables
deterministic rendering.

This means that the same scene rendered with the same settings will
result in a (near) identical image every time so tracking down bugs is
much easier.